Move debug UI framework to engine.
Also adds traits to simplify adding debug windows.
