+++ rs/Cargo.toml
@@ -11,7 +11,7 @@ members = ["tools/*", "game-dylib", "engine", "engine-dylib", "engine/macros"]
 
 [features]
 dylib = ["bevy_dylib", "sond-has-engine-dylib"]
-debugging = ["bevy-inspector-egui", "egui_plot"]
+debugging = ["engine/debugging", "bevy-inspector-egui", "egui_plot"]
 vis_test = ["testing", "render", "engine/vis_test"]
 testing = ["colored", "engine/testing"]
 render = ["engine/render"]
