+++ rs/engine/src/lib.rs
@@ -1,5 +1,6 @@
 #![warn(unused_crate_dependencies)]
 
+use bevy::app::{App, Plugin};
 #[allow(unused_imports, clippy::single_component_path_imports)]
 #[cfg(all(debug_assertions, not(target_arch = "wasm32")))]
 use bevy_dylib;
@@ -10,4 +11,14 @@ pub mod nav;
 pub mod offloading;
 pub mod planet;
 pub mod settings;
+pub mod ui;
 pub mod util;
+
+#[derive(Debug)]
+pub struct EnginePlugin;
+
+impl Plugin for EnginePlugin {
+	fn build(&self, app: &mut App) {
+		app.add_plugins(ui::UiPlugin);
+	}
+}
