+++ rs/src/ui.rs
@@ -1,6 +1,7 @@
-use crate::util::IntoFnPlugin;
 use bevy::prelude::*;
 use enum_components::EnumComponent;
+
+use crate::util::IntoFnPlugin;
 
 #[cfg(feature = "debugging")]
 pub mod dbg_ui;
@@ -12,8 +13,6 @@ pub fn plugin(app: &mut App) -> &mut App {
 	app.add_plugins(dbg_ui::plugin.plugfn());
 
 	app.add_plugins(game_ui::plugin.plugfn())
-		.init_resource::<UiHovered>()
-	// .add_plugins(pause_menu::plugin.plugfn())
 }
 
 #[derive(Debug, EnumComponent)]
@@ -21,8 +20,3 @@ pub enum GameMenu {
 	MainMenu,
 	PauseMenu,
 }
-
-/// Keeps track of whether a UI element is hovered over so that clicking
-/// does not grab the mouse if so.
-#[derive(Resource, Debug, Default, Copy, Clone, Deref, DerefMut, Reflect)]
-pub struct UiHovered(bool);
